Fixed hang when clicking on near the bottom of an Edit field.
If the last line of text in an edit field contains text that descends
before the baseline, clicking on the text near the bottom would cause
PickRow to assign start_it to the end of the string, and end_it to one
character before the end of the string. Having end_it < start_it caused
an infinite loop in PickColumn's std::upper_bound call.

With this fix, clicking that spot causes the carat to be placed at the
end of the offending line. There might be a better fix for this but
this one prevents the crash at least.

Tested: On Mac. Click near the bottom of an edit field containing text
that descends below the baseline, such as a comma or certain letters.
Clicking around the letter with the descender would cause a crash before
this commit.

Change-Id: I0ac69d045da1346c9cb27189cb7826cd66159995