Get compiling on VS2010.
VS2010 doesn't have great support for `auto`, especially
of aligned types.

The power-of-2 function in mathfu is intended for float
types, but we were using it here for int types. VS2010
didn't like it because it didn't know which pow() to
use. I've added a local version that does the casting.

Change-Id: I450c99d7f4f8bfc919b692c60ca6be1fdacd69e3
Tested: Compiles on VS2010 and Linux.