Adding the starter files for FlatUI API Reference.
This commit includes the necessary files to get doxygen to generate
the API Reference.

It also includes doxygen style comments for `flatui_common.h` and
`flatui_util.h`.

Bug: b/24940399
Tested: Tested on Ubuntu with generate_docs.py and viewed on Chrome.
I also built the library on Ubuntu to ensure that I did not
accidentally break the build.

Change-Id: I912db5d5fc501f89759aa620647e276f0cdefa96