Improve precision of glyph bounds calculation.
Previously we returned only one AABB, which is fine if the
range of glyphs is contiguous, but if the glyphs span multiple
lines, then the returned box will be too big.

The new version returns a *vector* of AABBs, one for each line.

Tested: Tested with various label sizes, over multiple lines, on OSX.
Still requires unit tests.

Change-Id: Ie4e6c764a06008f4c7cd605167c78b62b1561d2b