Add links to the FontBuffer so that they will be cached.
We were previously not filling in the `links` parameter when
returning the FontBuffer from the cache, in calls to GetHtmlBuffer().

The fix is to move the links into the FontBuffer class itself,
so that they get cached with the rest of FontBuffer. Also creates
a cleaner interface. When links aren't used, the overhead memory cost
is insignificant.

Change-Id: Ifeeab6f15dffe0206145133139d3713dbceafcc5
Tested: Links are returned properly, on OSX.
Bug: 32004805