Change FlatUI so that animations are mapped by HashedId instead of strings.
Change the animations map inside of PersistentState to map Anims to HashedIds rather than strings.

Add implementations of StartAnimation and Animatable that take in a HashedId instead of string.

Change FindAnim to take in a HashedId rather than a string.

Change the string versions of Animatable and StartAnimation to hash the input string id and call the HashedId implementations.
PiperOrigin-RevId: 141452870
Change-Id: If7958a954e8b9e2b73bd754e7c3f034ce7b747a5