Create function for AnimationTimeRemaining.
Cloned from CL 141121735 by 'g4 patch'.
Original change by laijess@laijess:flatuianimdemo:73:citc on 2016/12/05 17:13:23.

Change FlatUI so that animations are mapped by HashedId instead of strings.

Change the animations map inside of PersistentState to map Anims to HashedIds rather than strings.

Add implementations of StartAnimation and Animatable that take in a HashedId instead of string.

Change FindAnim to take in a HashedId rather than a string.

Change the string versions of Animatable and StartAnimation to hash the input string id and call the HashedId implementations.
PiperOrigin-RevId: 141467244
Change-Id: I479f91d70ba574b25e052df026c950f7e13992cc