Move test/flatuitest.cpp to sample/advanced/flatuiadvanced.cpp.
Cmake was having a problem building flatuitest.cpp because its
build commands were in the same test/CMakeLists.txt file as the
gtest commands. The gtest commands set Visual Studio common library
option to be incompatible with the rest of the code (static link
instead of dynamic link).

The fix is to move flatuitest.cpp out of the test/CMakeLists.txt
file. This integration test fits best with the samples, so I've moved
it over there.

I've also removed the unit_tests directory, since it's now redundant.

Tested: Samples and tests build and run on Linux and Mac.
PiperOrigin-RevId: 141509462
Change-Id: I0eae0f49f84c84fc1cfeb7524163ccbc05388651