Add Sprite capability to FlatUI animation.
Create functions AddSprite and DrawSprite that allow sprites to
be created inside of a user's app. Additionally, create an internal
Sprite struct that allows the animation system to keep store, handle,
and delete sprites as necessary.

These sprites will allow FlatUI to have "fire and forget"
functionality. An event can trigger a sprite to appear on the screen
and remain on screen until its animation is complete or be deleted if
the program never draws it. Once its animation is complete, the
FlatUI animation system will delete it.
PiperOrigin-RevId: 142040887
Change-Id: I75c77aa8b49f9b2ca988a2f4df04e1d809a2f3ec