Fix Android issue on CorgUI.
Previously, there was a check in the top while loop to see if the back button
on an Android device had been hit. However, this field would not be updated
until after input's advance frame was called. This was leading to seg faults as
the app tried to keep running after the back had already been hit.
PiperOrigin-RevId: 144393016
Change-Id: I2951293cee2bf3838e7d78785eb9a20b9515d52c