Fix Margin issue on CorgUI buttons.
Previously, margins were used to get the textures for the CorgUI buttons to align
correctly. This made it so that additional white space was clickable.
Now, the corgi body and head are changed to two different groups. This allows for the use of offsets for the groups rather than margins for the buttons.
PiperOrigin-RevId: 144452631
Change-Id: I2f3ff21889912038dba88a6b9e1bb36543cf4d6c