Give all backends the same external API.
We abstract the handle types for all backends so that we can present
a uniform interface to the user, regardless of which backend is used.

This is nice from a usage point of view. All libraries present the same interface.

It's also required for when we allow OpenGL or Vulkan to be selected
at runtime, instead of compile time.

This change lets us remove the FPLBASE_OPENGL define, since the API is now
backend agnostic.
PiperOrigin-RevId: 144479099
Change-Id: I7b7a9c28e48841ecf2ce72116531085b742f6f8f