Fix positioning and coloring of score label on CorgUI.
Play testing revealed that users felt the score label on CorgUI was located in
an awkward position. Feedback from teammates suggested that placing the score
in the upper left corner, similar to how it's done in many other games, would
be better.
Testers also felt the colors that the score label flashed when hitting a 100
point mark were off.
PiperOrigin-RevId: 144485168
Change-Id: I091e9a4f51d59df9b788048ba3bef0170dd1a1ae