Make Texture::UpdateTexture() non-static.
This function uses member data on Vulkan.

Also made Texture::CreateTexture() private. This function returns a handle, which is insufficient in Vulkan, where the TextureImpl state is also set. If we want this function in the external interface, we can (alternatively) initialize a Texture instead of just returning a handle.
PiperOrigin-RevId: 144790296
Change-Id: I1b8ffc3bdf273120e8d86fe8721310c07eea6ff5