Enlarge the corgi in CorgUI.
Make the corgi bigger to fill up more of the screen space on CorgUI.
Achieve by adding variable, kCorgiScale, which scales up the corgi.
This allows for less manual modification of each of the corgi's
individual assets, as the corgi is separated into multiple parts.

For variables that are not sizes or offsets, manual changes
may still need to be made to look correct. Manually modify
the start y position of the sprites to be above the corgi with
new, larger size.
PiperOrigin-RevId: 145477078
Change-Id: Ib5078035ec2f762e226bccebb44a8c02e013ab23