Fix hardware scaler related issue for apps running on Android.
Add ability to choose a window mode. Modes include
kWindowModeWindowedDesktopNative and kWindowModeWindowedDesktopScaled, which
both use a user input window_size. Also included are
kWindowModeFullscreenScaled, which enables fullscreen, and
kWindowModeFullscreenNative, which enables fullscreen and disables the hardware
scaler.

The only difference between kWindowModeWindowedDesktopNative and
kWindowModeWindowedDesktopScaled is what the app will default to if running on a
mobile platform and set to one of these. If it's set to
kWindowModeWindowedDesktopNative, it will set to kWindowModeFullscreenNative.
Likewise, if set to kWindowModeWindowedDesktopScaled, it will set to
kWindowModeFullscreenScaled.

This is particularly relevant for Android apps that use portrait mode,
as the scaler will cause the app to be rendered incorrectly in the y dimension
depending on the window size.
PiperOrigin-RevId: 145488702
Change-Id: I14b3835e65ceac942bfea9fdc3b556daa949a739