Fix flashing text in CorgUI.
A bug in FlatUI resulted in a label not being found in the render phase. This
was because labels were being hashed by their text, resulting in the label
not being found if its text value changed.

Add optional hash parameter that allows the user to pass in their own hash
for the label.
PiperOrigin-RevId: 145832373
Change-Id: I59b73261bb9f235e55ca6769fccf8b2123bec274