Fixing Zooshi Android build issues, mostly from Windows.
Android makefile's abspath function is bugged in Windows, with no good
workaround (it does not return a valid path). While it would make the
paths being used make more sense, using the absolute path is
unnecessary, so removing.
Also removing an extra line in fplutil, and removing the Cardboard
copy from the build gradle, as that was meant to only be in gerrit.

Note, this alone does not fix it entirely on Windows (it is not a fan
of splitting assets between assets and built_assets), but since that
split is needed in google3, it will need to be done in gerrit only
instead.

Tested with Android build on Windows and google3 (Linux).

PiperOrigin-RevId: 150810167
Change-Id: Ia198d96ba14c292ea4b5698446fb09e13372f6b2