Specified kFormat8888 on all relevant textures in samples.
By default, LoadTexture() will use kFormatAuto if not otherwise specified. This results in the image not using 8888, which is generally what is desired since most of our assets use an alpha channel. Although this is a bigger issue (ie: kFormatAuto should probably behave differently, and I have filed a bug report as such), as a temporary fix I've made all of our samples specify to use kFormat8888 where appropriate.

As an aside I fixed most of the linter errors (moving includes around, inserting a space after commas and //'s).

PiperOrigin-RevId: 160974960
Change-Id: I871081b9e2059fbf5cbcbe2d4e27053e63189cda